+++ R/upsert.R
@@ -6,7 +6,7 @@
 #' @param where_cols The columns to use for WHERE clause
 #' @param batch_size The number of records to upsert in a single statement (defaults to all)
 #' @param returning Columns to return
-#' @param ignore_dups Ignore duplicate rows
+#' @param skip_existing Skip existing rows
 #' @export
 #' @examples \dontrun{
 #'
@@ -17,7 +17,7 @@
 #' records <- data.frame(id=c(3, 4), temperature=c(20, 25))
 #' dbxUpsert(db, table, records, where_cols=c("id"))
 #' }
-dbxUpsert <- function(conn, table, records, where_cols, batch_size=NULL, returning=NULL, ignore_dups=FALSE) {
+dbxUpsert <- function(conn, table, records, where_cols, batch_size=NULL, returning=NULL, skip_existing=FALSE) {
   cols <- colnames(records)
 
   if (!identical(intersect(cols, where_cols), where_cols)) {
@@ -36,7 +36,7 @@ dbxUpsert <- function(conn, table, records, where_cols, batch_size=NULL, returni
   inBatches(records, batch_size, function(batch) {
     if (isMySQL(conn)) {
       sql <- insertClause(conn, table, batch)
-      if (ignore_dups) {
+      if (skip_existing) {
         # do not use INSERT IGNORE
         # https://stackoverflow.com/questions/2366813/on-duplicate-key-ignore
         set_sql <- upsertSetClause(quoted_where_cols)
@@ -49,7 +49,7 @@ dbxUpsert <- function(conn, table, records, where_cols, batch_size=NULL, returni
       conflict_target <- colsClause(quoted_where_cols)
       sql <- insertClause(conn, table, batch)
       sql <- paste0(sql, " ON CONFLICT (", conflict_target, ") DO")
-      if (ignore_dups) {
+      if (skip_existing) {
         sql <- paste(sql, "NOTHING")
       } else {
         set_sql <- upsertSetClausePostgres(quoted_update_cols)