+++ tests/testthat/helper-run-upsert-tests.R
@@ -29,14 +29,14 @@ runUpsertTests <- function(db, redshift=FALSE) {
     expect_equal(res$city, c("San Francisco", "Boston", NA))
   })
 
-  test_that("upsert ignore_dups works", {
+  test_that("upsert skip_existing works", {
     skip_if_not(upsertSupported())
 
     events <- data.frame(id=c(1, 2), city=c("San Francisco", "Boston"), stringsAsFactors=FALSE)
     dbxInsert(db, "events", events)
 
     upsert_events <- data.frame(id=c(2, 3), city=c("Chicago", "New York"))
-    dbxUpsert(db, "events", upsert_events, where_cols=c("id"), ignore_dups=TRUE)
+    dbxUpsert(db, "events", upsert_events, where_cols=c("id"), skip_existing=TRUE)
 
     res <- dbxSelect(db, "SELECT city FROM events ORDER BY id")
     expect_equal(res$city, c("San Francisco", "Boston", "New York"))